wip: testing: Add wrapper script to build certificates in root image
This does not modify the root image but uses the strongSwan version
installed there (avoids build dependencies on version installed on the
host to use pki to generate all the keys and certificates).

wip: We can perhaps rename the scripts so this is the actual main script
and the other only the helper used inside the chroot.